refactor: modularize neovim configuration
- Move all vim options to lua/options.lua

- Create separate plugin configuration files in lua/plugins/

- Organize plugins by functionality

- Preserve all original functionality and settings

- No functional changes, only organizational improvements
