+++ lua/plugins/init.lua
@@ -1,0 +1,196 @@
+return {
+    -- Git related plugins
+    'tpope/vim-fugitive',
+    'tpope/vim-rhubarb',
+    'tpope/vim-sleuth', -- Detect tabstop and shiftwidth automatically
+    
+    -- Lazygit integration
+    {
+        "kdheepak/lazygit.nvim",
+        cmd = {
+            "LazyGit",
+            "LazyGitConfig",
+            "LazyGitCurrentFile",
+            "LazyGitFilter",
+            "LazyGitFilterCurrentFile",
+        },
+        dependencies = {
+            "nvim-lua/plenary.nvim",
+        },
+        keys = {
+            { "<leader>lg", "<cmd>LazyGit<cr>", desc = "LazyGit" },
+            { "<leader>lc", "<cmd>LazyGitCurrentFile<cr>", desc = "LazyGit Current File" }
+        }
+    },
+
+    -- Tmux navigation
+    {
+        "christoomey/vim-tmux-navigator",
+        cmd = {
+            "TmuxNavigateLeft",
+            "TmuxNavigateDown",
+            "TmuxNavigateUp",
+            "TmuxNavigateRight",
+            "TmuxNavigatePrevious",
+        },
+        keys = {
+            { "<c-h>", "<cmd><C-U>TmuxNavigateLeft<cr>" },
+            { "<c-j>", "<cmd><C-U>TmuxNavigateDown<cr>" },
+            { "<c-k>", "<cmd><C-U>TmuxNavigateUp<cr>" },
+            { "<c-l>", "<cmd><C-U>TmuxNavigateRight<cr>" },
+            { "<c-\\>", "<cmd><C-U>TmuxNavigatePrevious<cr>" },
+        },
+    },
+
+    -- Surround plugin
+    {
+        "kylechui/nvim-surround",
+        version = "*",
+        event = "VeryLazy",
+        config = function()
+            require("nvim-surround").setup({})
+        end
+    },
+
+    -- Useful plugin to show you pending keybinds
+    { 'folke/which-key.nvim', opts = {} },
+
+    -- "gc" to comment visual regions/lines
+    { 'numToStr/Comment.nvim', opts = {} },
+
+    -- Adds git related signs to the gutter, as well as utilities for managing changes
+    {
+        'lewis6991/gitsigns.nvim',
+        opts = {
+            signs = {
+                add = { text = '+' },
+                change = { text = '~' },
+                delete = { text = '_' },
+                topdelete = { text = '‾' },
+                changedelete = { text = '~' },
+            },
+            on_attach = function(bufnr)
+                local gs = package.loaded.gitsigns
+
+                local function map(mode, l, r, opts)
+                    opts = opts or {}
+                    opts.buffer = bufnr
+                    vim.keymap.set(mode, l, r, opts)
+                end
+
+                map({ 'n', 'v' }, ']c', function()
+                    if vim.wo.diff then
+                        return ']c'
+                    end
+                    vim.schedule(function()
+                        gs.next_hunk()
+                    end)
+                    return '<Ignore>'
+                end, { expr = true, desc = 'Jump to next hunk' })
+
+                map({ 'n', 'v' }, '[c', function()
+                    if vim.wo.diff then
+                        return '[c'
+                    end
+                    vim.schedule(function()
+                        gs.prev_hunk()
+                    end)
+                    return '<Ignore>'
+                end, { expr = true, desc = 'Jump to previous hunk' })
+
+                -- Actions
+                -- visual mode
+                map('v', '<leader>hs', function()
+                    gs.stage_hunk { vim.fn.line '.', vim.fn.line 'v' }
+                end, { desc = 'stage git hunk' })
+                map('v', '<leader>hr', function()
+                    gs.reset_hunk { vim.fn.line '.', vim.fn.line 'v' }
+                end, { desc = 'reset git hunk' })
+                -- normal mode
+                map('n', '<leader>hs', gs.stage_hunk, { desc = 'git stage hunk' })
+                map('n', '<leader>hr', gs.reset_hunk, { desc = 'git reset hunk' })
+                map('n', '<leader>hS', gs.stage_buffer, { desc = 'git Stage buffer' })
+                map('n', '<leader>hu', gs.undo_stage_hunk, { desc = 'undo stage hunk' })
+                map('n', '<leader>hR', gs.reset_buffer, { desc = 'git Reset buffer' })
+                map('n', '<leader>hp', gs.preview_hunk, { desc = 'preview git hunk' })
+                map('n', '<leader>hb', function()
+                    gs.blame_line { full = false }
+                end, { desc = 'git blame line' })
+                map('n', '<leader>hd', gs.diffthis, { desc = 'git diff against index' })
+                map('n', '<leader>hD', function()
+                    gs.diffthis '~'
+                end, { desc = 'git diff against last commit' })
+
+                -- Toggles
+                map('n', '<leader>tb', gs.toggle_current_line_blame, { desc = 'toggle git blame line' })
+                map('n', '<leader>td', gs.toggle_deleted, { desc = 'toggle git show deleted' })
+
+                -- Text object
+                map({ 'o', 'x' }, 'ih', ':<C-U>Gitsigns select_hunk<CR>', { desc = 'select git hunk' })
+            end,
+        },
+    },
+
+    -- Theme
+    {
+        "folke/tokyonight.nvim",
+        lazy = false,
+        priority = 1000,
+        opts = {},
+    },
+
+    -- Harpoon for quick file navigation
+    {
+        "ThePrimeagen/harpoon",
+        branch = "harpoon2",
+        dependencies = { "nvim-lua/plenary.nvim" }
+    },
+
+    -- GitHub Copilot
+    { "github/copilot.vim" },
+    {
+        "CopilotC-Nvim/CopilotChat.nvim",
+        dependencies = {
+            { "github/copilot.vim" },
+            { "nvim-lua/plenary.nvim", branch = "master" },
+        },
+        build = "make tiktoken",
+        opts = {},
+    },
+
+    -- UI enhancements
+    { 'stevearc/dressing.nvim', opts = {} },
+    {
+        'windwp/nvim-autopairs',
+        event = "InsertEnter",
+        config = true
+    },
+    {
+        'nvim-lualine/lualine.nvim',
+        opts = {
+            options = {
+                icons_enabled = false,
+                theme = 'onedark',
+                component_separators = '|',
+                section_separators = '',
+            },
+        },
+    },
+
+    -- Indentation guides
+    {
+        "lukas-reineke/indent-blankline.nvim",
+        main = "ibl",
+        opts = {},
+    },
+
+    -- Add the rest of your plugins here
+    require('plugins.telescope'),
+    require('plugins.treesitter'),
+    require('plugins.lsp'),
+    require('plugins.completion'),
+    require('plugins.harpoon'),
+    require('plugins.oil'),
+    require('plugins.copilot'),
+    require('plugins.dap'),
+}
