+++ lua/plugins/treesitter.lua
@@ -1,0 +1,89 @@
+return {
+    'nvim-treesitter/nvim-treesitter',
+    dependencies = {
+        'nvim-treesitter/nvim-treesitter-textobjects',
+        'nvim-treesitter/nvim-treesitter-context',
+    },
+    build = ':TSUpdate',
+    config = function()
+        -- Defer Treesitter setup after first render to improve startup time of 'nvim {filename}'
+        vim.defer_fn(function()
+            require('nvim-treesitter.configs').setup {
+                ensure_installed = { 'go', 'lua', 'python', 'tsx', 'javascript', 'typescript', 'bash', 'pug' },
+                auto_install = true,
+                highlight = { enable = true },
+                indent = { enable = false },
+                incremental_selection = {
+                    enable = true,
+                    keymaps = {
+                        init_selection = '<c-space>',
+                        node_incremental = '<c-space>',
+                        scope_incremental = '<c-s>',
+                        node_decremental = '<M-space>',
+                    },
+                },
+                textobjects = {
+                    select = {
+                        enable = true,
+                        lookahead = true,
+                        keymaps = {
+                            ['aa'] = '@parameter.outer',
+                            ['ia'] = '@parameter.inner',
+                            ['af'] = '@function.outer',
+                            ['if'] = '@function.inner',
+                            ['ac'] = '@class.outer',
+                            ['ic'] = '@class.inner',
+                        },
+                    },
+                    move = {
+                        enable = true,
+                        set_jumps = true,
+                        goto_next_start = {
+                            [']m'] = '@function.outer',
+                            [']]'] = '@class.outer',
+                        },
+                        goto_next_end = {
+                            [']M'] = '@function.outer',
+                            [']['] = '@class.outer',
+                        },
+                        goto_previous_start = {
+                            ['[m'] = '@function.outer',
+                            ['[['] = '@class.outer',
+                        },
+                        goto_previous_end = {
+                            ['[M'] = '@function.outer',
+                            ['[]'] = '@class.outer',
+                        },
+                    },
+                    swap = {
+                        enable = true,
+                        swap_next = {
+                            ['<leader>a'] = '@parameter.inner',
+                        },
+                        swap_previous = {
+                            ['<leader>A'] = '@parameter.inner',
+                        },
+                    },
+                },
+                fold = {
+                    enable = true,
+                    disable = {},
+                },
+            }
+        end, 0)
+
+        -- Setup treesitter-context
+        require('treesitter-context').setup {
+            enable = true,
+            max_lines = 0,
+            min_window_height = 0,
+            line_numbers = true,
+            multiline_threshold = 20,
+            trim_scope = 'outer',
+            mode = 'cursor',
+            separator = nil,
+            zindex = 20,
+            on_attach = nil,
+        }
+    end,
+}
